refactor: split rates policy snapshot builder
